version 1.1-1

+++ src/pred_rpart.c
@@ -28,10 +28,10 @@
 #include "rpart.h"
 #include "rpartproto.h"
 
-void pred_rpart(long *dimx,	long *nnode, 	long *nsplit, 	long *dimc, 
-		long *nnum,  	long *nodes2,   long *vnum,     double *split2,
-		long *csplit2,  long *usesur,   double *xdata2, 
-		long *xmiss2,   long *where)
+void pred_rpart(int *dimx,	int *nnode, 	int *nsplit, 	int *dimc, 
+		int *nnum,  	int *nodes2,   int *vnum,     double *split2,
+		int *csplit2,  int *usesur,   double *xdata2, 
+		int *xmiss2,   int *where)
     {
     int i,j;
     int n;
@@ -40,9 +40,9 @@ void pred_rpart(long *dimx,	long *nnode, 	long *nsplit, 	long *dimc,
     int lcount, rcount;
     int npos;
     double temp;
-    long   *nodes[4];
+    int   *nodes[4];
     double *split[4];
-    long   **csplit,
+    int   **csplit,
 	   **xmiss;
     double **xdata;
 
@@ -53,10 +53,10 @@ void pred_rpart(long *dimx,	long *nnode, 	long *nsplit, 	long *dimc,
 	}
 
     if (dimc[1] > 0) {
-	csplit = (long **)  ALLOC((int)dimc[1], sizeof(int*));
+	csplit = (int **)  ALLOC((int)dimc[1], sizeof(int*));
 	for (i=0; i<dimc[1]; i++)  csplit[i] = &(csplit2[i * dimc[0]]);
 	}    
-    xmiss =  (long **)  ALLOC((int)dimx[1], sizeof(int*));
+    xmiss =  (int **)  ALLOC((int)dimx[1], sizeof(int*));
     xdata = (double **) ALLOC((int)dimx[1], sizeof(double*));
     for (i=0; i<dimx[1]; i++) {
 	xmiss[i] = &(xmiss2[i * dimx[0]]);
